update deploy script to be able to do a restart only as well

+++ deploy.sh
@@ -1,8 +1,19 @@
 #!/usr/bin/bash
 export PATH=$PATH:/home/gabor/docker-compose/bin
 
-git pull
-docker-compose build
-docker-compose stop --time 0
-docker-compose up --detach --time 0 --remove-orphans
-docker image prune -f
+if [ "$1" == "deploy" ];
+then
+  echo "deploy"
+  git pull
+  docker-compose build
+fi
+
+if [ "$1" == "deploy" ] || [ "$1" == "restart" ];
+then
+  echo "restart"
+  docker-compose stop --time 0
+  docker-compose up --detach --time 0 --remove-orphans
+  docker image prune -f
+else
+  echo "Usage: $0 [deploy|restart]"
+fi
